chg:All forexconnect common_samples moved to jgtfxcommon.py
