+++ jgtpy/_make_jgtfxcommon.sh
@@ -1,0 +1,7 @@
+for py in OrderMonitor.py BatchOrderMonitor.py OrderMonitorNetting.py TableListenerContainer.py common.py 
+do
+    echo " # $py" >> jgtfxcommon.py
+    cat common_samples/$py >> jgtfxcommon.py
+
+    echo "#------------------------#" >> jgtfxcommon.py
+done
