--- jgtpy/common_samples/__init__.py
@@ -1,23 +1,0 @@
-# Copyright 2019 Gehtsoft USA LLC
-
-# Licensed under the license derived from the Apache License, Version 2.0 (the "License"); 
-# you may not use this file except in compliance with the License.
-
-# You may obtain a copy of the License at
-
-# http://fxcodebase.com/licenses/open-source/license.html
-
-# Unless required by applicable law or agreed to in writing, software 
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from common_samples.BatchOrderMonitor import BatchOrderMonitor
-from common_samples.OrderMonitor import OrderMonitor
-from common_samples.OrderMonitorNetting import OrderMonitorNetting
-from common_samples.TableListenerContainer import TableListenerContainer
-from common_samples.common import add_main_arguments, add_instrument_timeframe_arguments, \
-    add_candle_open_price_mode_argument, add_direction_rate_lots_arguments, add_account_arguments, \
-    valid_datetime, add_date_arguments, add_report_date_arguments, add_max_bars_arguments, add_bars_arguments, \
-    print_exception, session_status_changed, diff_month, convert_timeframe_to_seconds
