+++ jgtpy/jgtfxc.test.py
@@ -7,12 +7,14 @@ status()
 connect()
 
 d=get_price_history('EUR/USD', 'H4')
+d.to_csv('../EURUSD_H4.csv')
 print(d)
 
 
 status()
 
 d=get_price_history('AUD/USD', 'H4')
+d.to_csv('../AUDUSD_H4.csv')
 print(d)
